fix: refactor handler transfers
